pyproject.toml
[build-system]
requires = ["setuptools>=65"]
build-backend = "setuptools.build_meta"

[project]
name = "diagram-ir"
version = "0.1.0"
description = "Automatic evaluation pipeline for educational math diagrams."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "Vishal Kumar" }]
dependencies = [
    "numpy>=1.23",
    "pandas>=1.5",
    "pydantic>=2.4",
    "python-dotenv>=1.0",
    "openai>=1.0",
    "together>=0.2",
    "tenacity>=8.2",
    "tiktoken>=0.5",
    "shapely>=2.0",
    "Pillow>=10.0",
    "tqdm>=4.66",
]

[project.optional-dependencies]
notebooks = [
    "aiofiles>=23.2",
    "matplotlib>=3.7",
    "scikit-learn>=1.2",
    "jinja2>=3.1",
]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29",
]

[tool.setuptools]
py-modules = [
    "IR_model",
    "geometry_engine",
    "evaluator",
    "backtranslation",
    "llm_judge",
    "llm_judge_schema",
    "make_confusion_table",
    "models",
    "prompts",
]
